Fix when rendering child jobs.

+++ unfurl/job.py
@@ -713,7 +713,10 @@ class Job(ConfigChange):
         if not WorkflowPlan:
             raise UnfurlError(f"unknown workflow: {joboptions.workflow}")  # type: ignore
 
-        rmtree(os.path.join(self.rootResource.base_dir, Folders.Planned))
+        if not self.parentJob:
+            # don't do this when running a nested job
+            # (planned was already removed and new tasks have already been rendered there)
+            rmtree(os.path.join(self.rootResource.base_dir, Folders.Planned))
         plan = WorkflowPlan(self.rootResource, self.manifest.tosca, joboptions)
         plan_requests = list(plan.execute_plan())
